refactor: split exercise routes into separate handlers
